requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
werkzeug==2.2.3
Flask-Login==0..6.2
Flask-Migrate==4.0.0
Flask-WTF==1.0.0
python-dotenv==0.19.0
pandas==1.3.3
openpyxl==3.0.9
flask-admin==1.6.0
celery==5.2.3
# svgnest==2.0.0  # Per il nesting
pymysql==1.1.0
email_validator==2.0.0
